move command out
